Add company update endpoint

+++ shared/api/src/main/java/com/userscompanies/model/Company.java
@@ -30,6 +30,6 @@ public class Company {
     String name;
 
     @Column
-    Long budget;
+    Double budget;
 
 }
